clean classifier analyses
